web_app/backend/requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
openai
llama-index-core
llama-index-embeddings-openai
llama-index-vector-stores-chroma
chromadb
